aegis-publisher/requirements.txt
fastapi
pydantic
google-cloud-bigquery
uvicorn
google-cloud-pubsub